Version 0.3.24
Improved the IndexDB handling. In a number of places users could have opened the IndexDB without creating the structure, this would have lead the the DB in place, but empty. This is an age old bug, but its now solved.

Also includes some fixes and improvements that have been sitting in the queue for a while.

Also includes a sample of how you could implement a click-to-dial solution.